+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.2.1] - 2020-01-16
 
 ## Added
 
@@ -19,13 +19,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 ## Changed
 
 - `Stream.Open()` and `Cursor.Next()` may now return `ErrStreamSealed`
-- `MemoryStream.ID()` will now panic if the `StreamID` field is empty
+- `MemoryStream.ID()` now panics if the `StreamID` field is empty
 - `MemoryStream.Append()` now panics if `Seal()` has been called
 - `MemoryStream.Append()` now panics if any of the given messages is `nil`
 
 ## Fixed
 
-- Fixed unconditional OCC failure in `ordered.Projector`
+- Fix unconditional OCC failure in `ordered.Projector`
 
 ## [0.2.0] - 2020-01-14
 
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 <!-- references -->
 [0.1.0]: https://github.com/dogmatiq/aperture/releases/tag/v0.1.0
 [0.2.0]: https://github.com/dogmatiq/aperture/releases/tag/v0.2.0
+[0.2.1]: https://github.com/dogmatiq/aperture/releases/tag/v0.2.1
 
 [Unreleased]: https://github.com/dogmatiq/aperture